requirements.txt
# مكتبات 
typer
rich
httpx
tldextract
rapidfuzz
python-whois
requests
dnspython>=2.0.0
python-dotenv>=1.0.0
rich>=10.0.0
requests>=2.0.0


